+++ server/__test__/router/feedback-snippet.test.js
@@ -1,6 +1,6 @@
 import request from 'supertest';
 
-import rejectSnippetFixture from '../__env__/fixtures/reject-snippet';
+import suggestSnippetFixture from '../__env__/fixtures/suggest-snippet';
 import { admin, editor, user } from '../__env__/fixtures/cookies';
 
 const app = process.app;
@@ -13,7 +13,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('200 and "success:true" if ROLE_ADMIN', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .set('Cookie', [adminCookie])
       .send()
@@ -21,7 +21,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
         if (err) return done(err);
 
         expect(res.status).toBe(200);
-        expect(res.body.data).toMatchObject(rejectSnippetFixture);
+        expect(res.body.data).toMatchObject(suggestSnippetFixture);
 
         return done();
       });
@@ -29,7 +29,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('200 if ROLE_EDITOR', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .set('Cookie', [editorCookie])
       .send()
@@ -43,7 +43,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('401 if no cookie', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .send()
       .expect(401)
@@ -52,7 +52,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('401 if role ROLE_USER', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .set('Cookie', [userCookie])
       .send()
@@ -61,7 +61,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   });
   test('400 if countries not passed', done => {
     return request(app)
-      .put('/api/feedback/snippet/id123/reject?categories[]=aml-authority')
+      .put('/api/feedback/snippet/id123/suggest?categories[]=aml-authority')
       .set('Cookie', [adminCookie])
       .send()
       .expect(400)
@@ -69,7 +69,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   });
   test('400 if categories not passed', done => {
     return request(app)
-      .put('/api/feedback/snippet/id123/reject?countries[]=DK')
+      .put('/api/feedback/snippet/id123/suggest?countries[]=DK')
       .set('Cookie', [adminCookie])
       .send()
       .expect(400)
@@ -87,7 +87,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('200 if valid suggestedCategories passed', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .set('Cookie', [adminCookie])
       .send({ suggestedCategories: ['id1'] })
@@ -97,7 +97,7 @@ describe('GET /feedback/snippet/:snippetId/:feedbackType', () => {
   test('400 if invalid suggestedCategories passed', done => {
     return request(app)
       .put(
-        '/api/feedback/snippet/id123/reject?countries[]=GB&categories[]=aml-authority'
+        '/api/feedback/snippet/id123/suggest?countries[]=GB&categories[]=aml-authority'
       )
       .set('Cookie', [adminCookie])
       .send({ suggestedCategories: '' })
