+++ server/src/router/feedback-snippet.route.js
@@ -1,13 +1,13 @@
 import boom from 'boom';
 
 export default ({ config, searchApiService }) => async (req, res, next) => {
-  const { snippetId, action } = req.params;
+  const { snippetId } = req.params;
   const { userId } = req.user || {};
   const { categories, countries } = req.query;
   const { suggestedCategories } = req.body;
 
   try {
-    const apiResponse = await searchApiService.rejectSnippet({
+    const apiResponse = await searchApiService.feedbackSnippet({
       snippetId,
       userId,
       categories,
@@ -17,7 +17,7 @@ export default ({ config, searchApiService }) => async (req, res, next) => {
 
     res.json({ data: apiResponse.data });
   } catch (err) {
-    console.error('searchApiService.rejectSnippet failed', { err });
-    return next(boom.forbidden('searchApiService.rejectSnippet failed'));
+    console.error('searchApiService.feedbackSnippet failed', { err });
+    return next(boom.forbidden('searchApiService.feedbackSnippet failed'));
   }
 };
