+++ server/src/validations/feedback-snippet.validation.js
@@ -1,6 +1,6 @@
 import joi from 'joi';
 
-import { FEEDBACK_REJECT } from '../constants/feedbackTypes';
+import { FEEDBACK_SUGGEST } from '../constants/feedbackTypes';
 
 import querySchema from './query.validation';
 
@@ -8,7 +8,7 @@ const feedbackSnippetSchema = {
   query: querySchema.query,
   params: joi.object().keys({
     snippetId: joi.string().required(),
-    action: joi.string().only([FEEDBACK_REJECT])
+    action: joi.string().only([FEEDBACK_SUGGEST])
   }),
   body: joi.object().keys({
     suggestedCategories: joi.array().items(joi.string())
